+++ sites/partners/src/listings/PaperListingForm/index.tsx
@@ -13,6 +13,10 @@ import {
   TimeFieldPeriod,
   Modal,
   AppearanceBorderType,
+  Tabs,
+  TabList,
+  Tab,
+  TabPanel,
   LatitudeLongitude,
 } from "@bloom-housing/ui-components"
 import { useForm, FormProvider } from "react-hook-form"
@@ -169,11 +173,8 @@ const defaults: FormListing = {
   waitlistMaxSize: null,
   isWaitlistOpen: null,
   waitlistOpenSpots: null,
-  whatToExpect: {
-    applicantsWillBeContacted: "",
-    allInfoWillBeVerified: "",
-    bePreparedIfChosen: "",
-  },
+  whatToExpect:
+    "Applicants will be contacted by the property agent in rank order until vacancies are filled. All of the information that you have provided will be verified and your eligibility confirmed. Your application will be removed from the waitlist if you have made any fraudulent statements. If we cannot verify a housing preference that you have claimed, you will not receive the preference but will not be otherwise penalized. Should your application be chosen, be prepared to fill out a more detailed application and provide required supporting documents.",
   units: [],
   accessibility: "",
   amenities: "",
@@ -192,6 +193,7 @@ const defaults: FormListing = {
   urlSlug: undefined,
   showWaitlist: false,
   reviewOrderType: null,
+  unitsSummary: [],
   unitsSummarized: {
     unitTypes: [],
     priorityTypes: [],
@@ -353,6 +355,7 @@ const ListingForm = ({ listing, editMode }: ListingFormProps) => {
 
   const { listingsService } = useContext(AuthContext)
 
+  const [tabIndex, setTabIndex] = useState(0)
   const [alert, setAlert] = useState<AlertErrorType | null>(null)
   const [loading, setLoading] = useState<boolean>(false)
   const [status, setStatus] = useState<ListingStatus>(null)
@@ -526,35 +529,77 @@ const ListingForm = ({ listing, editMode }: ListingFormProps) => {
 
                 <Form id="listing-form" onSubmit={handleSubmit(triggerSubmit, onError)}>
                   <div className="flex flex-row flex-wrap">
-                    <div className="info-card md:w-9/12">
-                      <ListingIntro />
-                      <ListingPhoto />
-                      <BuildingDetails
-                        listing={listing}
-                        setLatLong={setLatitudeLongitude}
-                        latLong={latLong}
-                        customMapPositionChosen={customMapPositionChosen}
-                        setCustomMapPositionChosen={setCustomMapPositionChosen}
-                      />
-                      <CommunityType listing={listing} />
-                      <Units
-                        units={units}
-                        setUnits={setUnits}
-                        disableUnitsAccordion={listing?.disableUnitsAccordion}
-                      />
-                      <Preferences preferences={preferences} setPreferences={setPreferences} />
-                      <AdditionalFees />
-                      <BuildingFeatures />
-                      <AdditionalEligibility />
-                      <AdditionalDetails />
-                      <RankingsAndResults listing={listing} />
-                      <LeasingAgent />
-                      <ApplicationAddress listing={listing} />
-                      <ApplicationDates
-                        listing={listing}
-                        openHouseEvents={openHouseEvents}
-                        setOpenHouseEvents={setOpenHouseEvents}
-                      />
+                    <div className="md:w-9/12 pb-24">
+                      <Tabs
+                        forceRenderTabPanel={true}
+                        selectedIndex={tabIndex}
+                        onSelect={(index) => setTabIndex(index)}
+                      >
+                        <TabList>
+                          <Tab>Listing Details</Tab>
+                          <Tab>Application Process</Tab>
+                        </TabList>
+                        <TabPanel>
+                          <ListingIntro />
+                          <ListingPhoto />
+                          <BuildingDetails
+                            listing={listing}
+                            setLatLong={setLatitudeLongitude}
+                            latLong={latLong}
+                            customMapPositionChosen={customMapPositionChosen}
+                            setCustomMapPositionChosen={setCustomMapPositionChosen}
+                          />
+                          <CommunityType listing={listing} />
+                          <Units
+                            units={units}
+                            setUnits={setUnits}
+                            disableUnitsAccordion={listing?.disableUnitsAccordion}
+                          />
+                          <Preferences preferences={preferences} setPreferences={setPreferences} />
+                          <AdditionalFees />
+                          <BuildingFeatures />
+                          <AdditionalEligibility />
+                          <AdditionalDetails />
+
+                          <div className="text-right -mr-8 -mt-8 relative" style={{ top: "7rem" }}>
+                            <Button
+                              type="button"
+                              icon="arrowForward"
+                              onClick={() => {
+                                setTabIndex(1)
+                                window.scrollTo({ top: 0, behavior: "smooth" })
+                              }}
+                            >
+                              Application Process
+                            </Button>
+                          </div>
+                        </TabPanel>
+                        <TabPanel>
+                          <RankingsAndResults listing={listing} />
+                          <LeasingAgent />
+                          <ApplicationAddress listing={listing} />
+                          <ApplicationDates
+                            listing={listing}
+                            openHouseEvents={openHouseEvents}
+                            setOpenHouseEvents={setOpenHouseEvents}
+                          />
+
+                          <div className="-ml-8 -mt-8 relative" style={{ top: "7rem" }}>
+                            <Button
+                              type="button"
+                              icon="arrowBack"
+                              iconPlacement="left"
+                              onClick={() => {
+                                setTabIndex(0)
+                                window.scrollTo({ top: 0, behavior: "smooth" })
+                              }}
+                            >
+                              Listing Details
+                            </Button>
+                          </div>
+                        </TabPanel>
+                      </Tabs>
+
                       {listing?.status === ListingStatus.closed && (
                         <LotteryResults
                           submitCallback={(data) => {